Potential Final Before Self-Grading
Lots of changes. Head is a little tired...but mostly adjusted the zip code margins to get the desired positioning and effect that I wanted. Various other stylings and positionings, fiddled with widths and margins when perhaps there was a much bettter option but this is technically my second time through this project.